Create Card_checker.cpp
After entering the credit card number, the system verifies its validity and determines whether it is a Visa, MasterCard, American Express, or another form of a card. It verifies the authenticity of the card and the type of credit card by performing some basic operations and validations. Credit card numbers entered by the user are generated along with a message indicating whether the card is valid or invalid. If the credit card number entered is valid, the credit card type will also be printed.

There are patterns in credit card numbers.

Credit card numbers must contain 13 to 16 digits. It must start with:

4 for Visa cards 
5 for Master cards 
37 for cards from American Express
6 for cards from Discover