I have added code for the Printing a Pyramid:
When you run this program and input the number of rows you want for your pyramid, it will generate a pyramid pattern with the specified number of rows and spaces to create a visually appealing pyramid shape. You can modify the character used for the pyramid by changing the printf("*") statement to another character if desired.