I have added the files of Prims Algorithm and kruskals Algorithm
Prim's algorithm is a greedy algorithm used to find the minimum spanning tree of a connected, undirected graph. This code reads the number of vertices, edges, and the edges' weights from the input and constructs the minimum spanning tree using Prim's algorithm. The minimum spanning tree is then printed as a list of edges with their weights.

Kruskal's algorithm is another algorithm for finding the minimum spanning tree of a connected, undirected graph.This code reads the number of vertices, edges, and the edges' weights from the input and constructs the minimum spanning tree using Kruskal's algorithm. The minimum spanning tree is then printed as a list of edges with their weights.