Feat  ✨ : Added Bubble Sort algorithm and profile data
This commit adds a JavaScript implementation of the Bubble Sort algorithm in the bubbleSort.js file. The algorithm sorts an array of numbers in ascending order. An example usage is provided in the comments.

Additionally, profile images have been added to the static/images/ folder 🖼️, and bio data files have been added to the content/participant/ folder 📝. Each participant's data is stored in their respective markdown file.